+++ packages/common/src/models/Game/GameColumn.ts
@@ -1,6 +1,0 @@
-export interface GameColumn {
-  main: number[];
-  plus?: number[];
-}
-
-export type GameColumnKeys = keyof GameColumn;