+++ packages/common/src/models/Ticket/Ticket.ts
@@ -1,0 +1,9 @@
+import { Column } from './Column';
+import { Game } from '../Game';
+
+export interface Ticket {
+  game: Game;
+  numbers: {
+    [colName: string]: Column;
+  };
+}