+++ packages/common/src/models/index.ts
@@ -5,3 +5,4 @@ export * from './Lottery';
 export * from './Match';
 export * from './Regular';
 export * from './Report';
+export * from './Ticket';